requirements.txt
google-generativeai
SpeechRecognition
gTTS
pygame
pyaudio